renamed graph_with_connexions to undirected_graph
